webcast_qa_ios-duanshanhu_dev_IOS/requirements.txt
shoots>=0.3.5
shoots-ios
shoots-uibase

byted-cv